Start a rewrite of ook in Python.
I realized something very important today - shell scripting is not limited to
shell scripting.

If we want to have a config file written in Python, we need to have the parsing
program also written in Python for it to be able to execute the config and use
the variables.  With shell scripting, we do this by `export`ing the variables
into child processes' environments.  But shell commands don't have to be
written in shell; they can be shell functions, or compiled binaries, or scripts
written in shell, Python, Ruby, etc.  That is to say, we can interoperate with
any number of languages, as long as they communicate using environment
variables and the standard input/output streams.

And there are good reasons to write this in Python instead of shell, namely
that bootstrapping is difficult.  For all the interaction we're going to have
with the web service, we'd need either some really ugly and error-prone awk, or
a proper JSON parser.  Now, if another script needs a JSON parser, you just
install it via ook, but we're trying to *create* ook - so it's easier to let
pip figure out the dependencies.